requirements.txt
# Production requirements for Cloud Run deployment
# These are also specified in pyproject.toml

django>=5.2.8
django-environ>=0.11.2
graphene-django>=3.2.3
psycopg>=3.2.12
requests>=2.32.5
responses>=0.24.0
httpx>=0.27.0
openai>=1.54.0
gunicorn>=23.0.0
whitenoise>=6.7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

